# docs/swagger.yaml
openapi: 3.0.3
info:
  title: UploadHaven API
  description: |
    UploadHaven is a modern, secure file sharing platform built with Next.js 15, TypeScript, and MongoDB. 
    This API provides comprehensive file management, real-time notifications, security scanning, and analytics capabilities.
    
    ## Features
    - 🚀 Secure file upload and sharing
    - 🔔 Real-time notifications via Server-Sent Events
    - 🔒 Malware scanning and security monitoring
    - 👥 User and admin management
    - 📊 Analytics and statistics
    - 🎯 Rate limiting and abuse prevention
    
    ## Authentication
    The API uses session-based authentication with cookies. Most endpoints require authentication.
    
  version: 2.0.0
  contact:
    name: UploadHaven Support
    url: https://github.com/yourusername/uploadhaven
    email: support@uploadhaven.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://uploadhaven.vercel.app
    description: Production server
security:
  - sessionAuth: []
...
paths:
...
  /api/user/files:
    get:
      tags: [Files, User]
      summary: Get user's files
      description: Get files uploaded by the authenticated user
      security:
        - sessionAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [uploadDate, filename, size, downloadCount]
            default: uploadDate
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: User's files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilesResponse'
        '401':
          description: Unauthorized
...
  /api/security/scan/file:
    post:
      tags: [Security]
      summary: Scan file for malware
      description: Scan a specific file for malware and security threats
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName]
              properties:
                fileName:
                  type: string
                  description: Filename to scan
                  example: "document.pdf"
      responses:
        '200':
          description: Scan completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'
        '404':
          description: File not found
        '500':
          description: Scan failed
  /api/security/scan/bulk:
    post:
      tags: [Security, Admin]
      summary: Bulk security scan
      description: Scan multiple files for security threats (admin only)
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filenames:
                  type: array
                  items:
                    type: string
                  description: Specific files to scan
                scanAll:
                  type: boolean
                  default: false
                  description: Scan all files in the system
      responses:
        '200':
          description: Bulk scan completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkScanResult'
        '403':
          description: Admin access required
  /api/security:
    get:
      tags: [Security, Admin]
      summary: Get security events
      description: Get security events and logs (admin only)
      security:
        - sessionAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: type
          in: query
          schema:
            type: string
          description: Filter by event type
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high]
          description: Filter by severity
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Security events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityEventsResponse'
        '403':
          description: Admin access required

  # Analytics Endpoints
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication using HTTP-only cookies
  schemas:
...
    PaginationData:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

    # Authentication Types
...
    FileInfo:
      type: object
      properties:
        _id:
          type: string
        filename:
          type: string
        shortUrl:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        uploadDate:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        downloadCount:
          type: integer
        isPasswordProtected:
          type: boolean
    FilesResponse:
      type: object
      properties:
        success:
          type: boolean
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileInfo'
        pagination:
          $ref: '#/components/schemas/PaginationData'

    # Notification Types
...
    ScanResult:
      type: object
      properties:
        fileName:
          type: string
        scanResult:
          type: object
          properties:
            isMalicious:
              type: boolean
            isSuspicious:
              type: boolean
            threatName:
              type: string
            source:
              type: string
              enum: [local, virustotal]
            scannedAt:
              type: string
              format: date-time
            details:
              type: string
        quotaStatus:
          type: object
          properties:
            remaining:
              type: integer
            used:
              type: integer
            limit:
              type: integer
            resetTime:
              type: string
              format: date-time
        scannedAt:
          type: string
          format: date-time
    BulkScanResult:
      type: object
      properties:
        success:
          type: boolean
        totalFiles:
          type: integer
        scannedFiles:
          type: integer
        threatsFound:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
              result:
                type: string
                enum: [clean, malicious, suspicious, error]
              threatName:
                type: string
              error:
                type: string
    SecurityEvent:
      type: object
      properties:
        _id:
          type: string
        type:
          type: string
        timestamp:
          type: string
          format: date-time
        ip:
          type: string
        details:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        userAgent:
          type: string
        filename:
          type: string
        userId:
          type: string
        metadata:
          type: object
          additionalProperties: true
    SecurityEventsResponse:
      type: object
      properties:
        success:
          type: boolean
        events:
          type: array
          items:
            $ref: '#/components/schemas/SecurityEvent'
        pagination:
          $ref: '#/components/schemas/PaginationData'

    # Analytics Types
...